backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
# Celery
celery==5.4.0
redis==5.2.0
python-multipart==0.0.6
# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.17.2
# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0


email-validator>=2.0.0